backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
openai>=1.0.0
python-multipart==0.0.9
pydantic==2.6.1
stripe==7.12.0
python-jose==3.3.0
reportlab==4.0.9
jinja2==3.1.3
